Move deprecated function placeholders in Bolt\Extensions to a trait
